additional housekeeping with refactoring in preparation of extending that application to service multiple clients
